Fixes: host list display for non-admin users, user list refresh after user save.

+++ zico-core/src/main/java/com/jitlogic/zico/core/services/HostGwtService.java
@@ -19,13 +19,20 @@ package com.jitlogic.zico.core.services;
 import com.jitlogic.zico.core.HostStore;
 import com.jitlogic.zico.core.HostStoreManager;
 import com.jitlogic.zico.core.User;
+import com.jitlogic.zico.core.UserContext;
 import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowCallbackHandler;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
 import javax.sql.DataSource;
 import java.io.IOException;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 
 @Singleton
@@ -33,16 +40,39 @@ public class HostGwtService {
 
     private HostStoreManager hsm;
     private JdbcTemplate jdbc;
+    private UserContext ctx;
 
     @Inject
-    public HostGwtService(DataSource ds, HostStoreManager hsm) {
+    public HostGwtService(DataSource ds, HostStoreManager hsm, UserContext ctx) {
         this.hsm = hsm;
         this.jdbc = new JdbcTemplate(ds);
+        this.ctx = ctx;
     }
 
 
     public List<HostStore> findAll() {
-        return hsm.list();
+        if (ctx.isInRole("ADMIN")) {
+            return hsm.list();
+        } else {
+            List<HostStore> lst = new ArrayList<HostStore>();
+            int uid = jdbc.queryForObject("select USER_ID from USERS where USER_NAME = ?", Integer.class, ctx.getUser());
+            final Set<Integer> allowedHosts = new HashSet<Integer>();
+
+            jdbc.query("select HOST_ID from USERS_HOSTS where USER_ID = ?", new RowCallbackHandler() {
+                @Override
+                public void processRow(ResultSet resultSet) throws SQLException {
+                    allowedHosts.add(resultSet.getInt("HOST_ID"));
+                }
+            }, uid);
+
+            for (HostStore host : hsm.list()) {
+                if (allowedHosts.contains(host.getId())) {
+                    lst.add(host);
+                }
+            }
+
+            return lst;
+        }
     }
 
 
@@ -52,11 +82,13 @@ public class HostGwtService {
 
 
     public void addUserHost(User user, HostStore host) {
+        ctx.checkAdmin();
         jdbc.update("insert ignore USERS_HOSTS (USER_ID, HOST_ID) values (?,?)", user.getId(), host.getId());
     }
 
 
     public void delUserHost(User user, HostStore host) {
+        ctx.checkAdmin();
         jdbc.update("delete from USERS_HOSTS where USER_ID = ? and HOST_ID = ?", user.getId(), host.getId());
     }
